services/post_service/requirements.txt
# fastapi + sqlalchemy
fastapi[standard]
uvicorn[standard]
SQLAlchemy==2.0.46
python-decouple==3.8
alembic==1.13.2

# pydantic
pydantic==2.8.2
pydantic-settings==2.4.0

# psycopg2
psycopg2-binary

# jwt
python-jose[crptography]
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart
jwt